explicit loss of precision to avoid compilation errors

+++ libraries/chain/util/sps_processor.cpp
@@ -339,7 +339,7 @@ void sps_processor::convert_funds( const block_notification& note )
     return;
   }
 
-  const auto to_convert_amount = ( HIVE_PROPOSAL_CONVERSION_RATE * treasury_account.balance.amount ) / HIVE_100_PERCENT;
+  const auto to_convert_amount = (uint64_t) HIVE_PROPOSAL_CONVERSION_RATE * treasury_account.balance.amount / HIVE_100_PERCENT;
   const auto to_convert = asset(to_convert_amount, HIVE_SYMBOL);
   elog( "total balance is  ${bal} to convert is ${conv}" , ("bal", treasury_account.balance.amount ) ("conv", to_convert) );
